Implemented routes for best code readability + some small adjustments in code
